requirements.txt
# Django Framework
Django>=4.2,<5.0
django-environ>=0.11.2
django-cors-headers>=4.3.1
django-csp>=3.8

# API
djangorestframework>=3.14.0
drf-spectacular>=0.27.0

# Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2

# Database (optional - SQLite default)
# psycopg2-binary>=2.9.9

# Production server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development
python-dotenv>=1.0.0
